# laravel-json-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TekManager Sistema de Gestão Integrado",
        "description": "Documentação geral da API de acesso aos dados",
        "contact": {
            "email": "cesar@berci.com.br"
        },
        "license": {
            "name": "TekSea ₢ 2024",
            "url": "http://www.teksea.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v2/rastreabilidades": {
            "get": {
                "summary": "Get all rastreabilidades",
                "description": "Tabela que armazena a rastreabilidade dos produtos",
                "operationId": "4d468a73d2e88bd43d12a21d4afed8eb",
                "responses": {
                    "200": {
                        "description": "Get all rastreabilidades"
                    }
                }
            },
            "post": {
                "summary": "Create a new rastreabilidade",
                "operationId": "8130f8abff9dd76befacfd038d790b42",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Identificador único para rastreabilidade de cada produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "numero_serie": {
                                        "description": "Número de série do produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_geracao": {
                                        "description": "Data de geração do produto ou registro",
                                        "type": "string",
                                        "example": "2022-01-01"
                                    },
                                    "responsavel_criacao": {
                                        "description": "Responsável pela criação do produto ou registro",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "cliente_id": {
                                        "description": "Identificador do cliente relacionado (FK para clientes)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "estoque_id": {
                                        "description": "Identificador do item no estoque relacionado (FK para estoque)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "pv": {
                                        "description": "PV associado ao produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "op": {
                                        "description": "Ordem de produção associada ao produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "codigo_net": {
                                        "description": "Código NET do produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ref_sistema": {
                                        "description": "Referência do sistema para o produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "produto": {
                                        "description": "Descrição do produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "obra_alocada": {
                                        "description": "Obra onde o produto foi alocado",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "n_fatura": {
                                        "description": "Número da fatura associada ao produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "data_enviado": {
                                        "description": "Data de envio do produto",
                                        "type": "string",
                                        "example": "2022-01-01"
                                    },
                                    "garantia_dias": {
                                        "description": "Número de dias de garantia do produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "expira_garantia": {
                                        "description": "Data de expiração da garantia",
                                        "type": "string",
                                        "example": "2022-01-01"
                                    },
                                    "status_garantia": {
                                        "description": "Status atual da garantia do produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "condicao_garantia": {
                                        "description": "Condições específicas da garantia",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "created_at": {
                                        "description": "Data de criação do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    },
                                    "updated_at": {
                                        "description": "Data de atualização do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new rastreabilidade"
                    }
                }
            }
        },
        "/api/v2/rastreabilidades/{id}": {
            "get": {
                "summary": "Get a rastreabilidade by ID",
                "operationId": "73db6b6708c0046d67ce7a09e8971b30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the rastreabilidade",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get a rastreabilidade by ID"
                    }
                }
            },
            "put": {
                "summary": "Update a rastreabilidade",
                "operationId": "b4d045060fe38253555ca5f972519596",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the rastreabilidade",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Identificador único para rastreabilidade de cada produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "numero_serie": {
                                        "description": "Número de série do produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_geracao": {
                                        "description": "Data de geração do produto ou registro",
                                        "type": "string",
                                        "example": "2022-01-01"
                                    },
                                    "responsavel_criacao": {
                                        "description": "Responsável pela criação do produto ou registro",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "cliente_id": {
                                        "description": "Identificador do cliente relacionado (FK para clientes)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "estoque_id": {
                                        "description": "Identificador do item no estoque relacionado (FK para estoque)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "pv": {
                                        "description": "PV associado ao produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "op": {
                                        "description": "Ordem de produção associada ao produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "codigo_net": {
                                        "description": "Código NET do produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ref_sistema": {
                                        "description": "Referência do sistema para o produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "produto": {
                                        "description": "Descrição do produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "obra_alocada": {
                                        "description": "Obra onde o produto foi alocado",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "n_fatura": {
                                        "description": "Número da fatura associada ao produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "data_enviado": {
                                        "description": "Data de envio do produto",
                                        "type": "string",
                                        "example": "2022-01-01"
                                    },
                                    "garantia_dias": {
                                        "description": "Número de dias de garantia do produto",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "expira_garantia": {
                                        "description": "Data de expiração da garantia",
                                        "type": "string",
                                        "example": "2022-01-01"
                                    },
                                    "status_garantia": {
                                        "description": "Status atual da garantia do produto",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "condicao_garantia": {
                                        "description": "Condições específicas da garantia",
                                        "type": "string",
                                        "example": "example"
                                    },
                                    "created_at": {
                                        "description": "Data de criação do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    },
                                    "updated_at": {
                                        "description": "Data de atualização do registro",
                                        "type": "string",
                                        "example": "2022-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update a rastreabilidade"
                    }
                }
            },
            "delete": {
                "summary": "Delete a rastreabilidade",
                "operationId": "e358570c983c4d6605eb42b94aaaf918",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the rastreabilidade",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete a rastreabilidade"
                    }
                }
            }
        },
...
    }
}